requirements.txt
django
django-crispy-forms
crispy-bootstrap5
python-magic
python-magic-bin
django-htmx